Adding WebGL build support, and some other features
Added WebGL build support thanks to "(Unity) Better Minimal WebGL Template" by Sean:
https://seansleblanc.itch.io/better-minimal-webgl-template

The above was also credited in the DocumentationMainPage.md which can be found in the Assets/Scripts directory.

User interface to change the Quality Setting was added to the Main Menu. This wasn't part of the project at the time, so I added it just now to provide basic graphics settings to the game.